Renamed app from AwesomeProject to Space Explorer...is working
