+++ android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.awesomeproject",
+    package = "com.julioecheverri.spaceexplorer",
 )
 
 android_resource(
     name = "res",
-    package = "com.awesomeproject",
+    package = "com.julioecheverri.spaceexplorer",
     res = "src/main/res",
 )
