+++ android/app/src/main/java/com/julioecheverri/spaceexplorer/MainActivity.java
@@ -1,4 +1,4 @@
-package com.awesomeproject;
+package com.julioecheverri.spaceexplorer;
 
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactActivityDelegate;
@@ -13,7 +13,7 @@ public class MainActivity extends ReactActivity {
    */
   @Override
   protected String getMainComponentName() {
-    return "AwesomeProject";
+    return "SpaceExplorer";
   }
 
   /**
