+++ android/app/src/main/java/com/julioecheverri/spaceexplorer/MainApplication.java
@@ -1,4 +1,4 @@
-package com.awesomeproject;
+package com.julioecheverri.spaceexplorer;
 
 import android.app.Application;
 import android.content.Context;
@@ -9,7 +9,7 @@ import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.config.ReactFeatureFlags;
 import com.facebook.soloader.SoLoader;
-import com.awesomeproject.newarchitecture.MainApplicationReactNativeHost;
+import com.julioecheverri.spaceexplorer.newarchitecture.MainApplicationReactNativeHost;
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
 
@@ -73,7 +73,7 @@ public class MainApplication extends Application implements ReactApplication {
          We use reflection here to pick up the class that initializes Flipper,
         since Flipper library is not available in release mode
         */
-        Class<?> aClass = Class.forName("com.awesomeproject.ReactNativeFlipper");
+        Class<?> aClass = Class.forName("com.julioecheverri.spaceexplorer.ReactNativeFlipper");
         aClass
             .getMethod("initializeFlipper", Context.class, ReactInstanceManager.class)
             .invoke(null, context, reactInstanceManager);
