+++ android/app/src/main/java/com/julioecheverri/spaceexplorer/newarchitecture/MainApplicationReactNativeHost.java
@@ -1,4 +1,4 @@
-package com.awesomeproject.newarchitecture;
+package com.julioecheverri.spaceexplorer.newarchitecture;
 
 import android.app.Application;
 import androidx.annotation.NonNull;
@@ -19,9 +19,9 @@ import com.facebook.react.fabric.CoreComponentsRegistry;
 import com.facebook.react.fabric.FabricJSIModuleProvider;
 import com.facebook.react.fabric.ReactNativeConfig;
 import com.facebook.react.uimanager.ViewManagerRegistry;
-import com.awesomeproject.BuildConfig;
-import com.awesomeproject.newarchitecture.components.MainComponentsRegistry;
-import com.awesomeproject.newarchitecture.modules.MainApplicationTurboModuleManagerDelegate;
+import com.julioecheverri.spaceexplorer.BuildConfig;
+import com.julioecheverri.spaceexplorer.newarchitecture.components.MainComponentsRegistry;
+import com.julioecheverri.spaceexplorer.newarchitecture.modules.MainApplicationTurboModuleManagerDelegate;
 import java.util.ArrayList;
 import java.util.List;
