+++ android/app/src/main/java/com/julioecheverri/spaceexplorer/newarchitecture/components/MainComponentsRegistry.java
@@ -1,4 +1,4 @@
-package com.awesomeproject.newarchitecture.components;
+package com.julioecheverri.spaceexplorer.newarchitecture.components;
 
 import com.facebook.jni.HybridData;
 import com.facebook.proguard.annotations.DoNotStrip;
