+++ android/app/src/main/jni/MainComponentsRegistry.h
@@ -13,7 +13,7 @@ class MainComponentsRegistry
  public:
   // Adapt it to the package you used for your Java class.
   constexpr static auto kJavaDescriptor =
-      "Lcom/awesomeproject/newarchitecture/components/MainComponentsRegistry;";
+      "Lcom/julioecheverri/spaceexplorer/newarchitecture/components/MainComponentsRegistry;";
 
   static void registerNatives();
