+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "AwesomeProject",
-  "displayName": "AwesomeProject"
+  "name": "SpaceExplorer",
+  "displayName": "Space Explorer"
 }
