+++ ios/SpaceExplorer.xcodeproj/xcshareddata/xcschemes/SpaceExplorer.xcscheme
@@ -15,9 +15,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-               BuildableName = "AwesomeProject.app"
-               BlueprintName = "AwesomeProject"
-               ReferencedContainer = "container:AwesomeProject.xcodeproj">
+               BuildableName = "SpaceExplorer.app"
+               BlueprintName = "SpaceExplorer"
+               ReferencedContainer = "container:SpaceExplorer.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -33,9 +33,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "00E356ED1AD99517003FC87E"
-               BuildableName = "AwesomeProjectTests.xctest"
-               BlueprintName = "AwesomeProjectTests"
-               ReferencedContainer = "container:AwesomeProject.xcodeproj">
+               BuildableName = "SpaceExplorerTests.xctest"
+               BlueprintName = "SpaceExplorerTests"
+               ReferencedContainer = "container:SpaceExplorer.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -55,9 +55,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "AwesomeProject.app"
-            BlueprintName = "AwesomeProject"
-            ReferencedContainer = "container:AwesomeProject.xcodeproj">
+            BuildableName = "SpaceExplorer.app"
+            BlueprintName = "SpaceExplorer"
+            ReferencedContainer = "container:SpaceExplorer.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </LaunchAction>
@@ -72,9 +72,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "13B07F861A680F5B00A75B9A"
-            BuildableName = "AwesomeProject.app"
-            BlueprintName = "AwesomeProject"
-            ReferencedContainer = "container:AwesomeProject.xcodeproj">
+            BuildableName = "SpaceExplorer.app"
+            BlueprintName = "SpaceExplorer"
+            ReferencedContainer = "container:SpaceExplorer.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
