+++ ios/SpaceExplorer/AppDelegate.mm
@@ -44,7 +44,7 @@ static NSString *const kRNConcurrentRoot = @"concurrentRoot";
 #endif
 
   NSDictionary *initProps = [self prepareInitialProps];
-  UIView *rootView = RCTAppSetupDefaultRootView(bridge, @"AwesomeProject", initProps);
+  UIView *rootView = RCTAppSetupDefaultRootView(bridge, @"SpaceExplorer", initProps);
 
   if (@available(iOS 13.0, *)) {
     rootView.backgroundColor = [UIColor systemBackgroundColor];
