+++ ios/SpaceExplorer/Images.xcassets/AppIcon.appiconset/Contents.json
