+++ ios/SpaceExplorerTests/Info.plist
